+++ proj_WolfoFixes/code/WolfoFixes.cs
@@ -11,7 +11,7 @@ namespace WolfoFixes
 {
 
     [BepInDependency("com.bepis.r2api")]
-    [BepInPlugin("com.Wolfo.WolfFixes", "WolfoBugFixes", "1.2.0")]
+    [BepInPlugin("com.Wolfo.WolfFixes", "WolfoBugFixes", "1.3.0")]
     [NetworkCompatibility(CompatibilityLevel.NoNeedForSync, VersionStrictness.DifferentModVersionsAreOk)]
     public class WolfFixes : BaseUnityPlugin
     {